A word guesser for Draw My Thing on the Minecraft server Mineplex.
This is a simple word guesser for Draw My Thing. It allows a user to
input a hint in the form _l___ where underscores are unknown letters in
a word. It has an auto-typing feature that uses the java class Robot to
type words quickly into Minecraft's chat. When you switch from
Minecraft's window to another program's window, it will automatically
stop auto-typing.